Make Mozilla injection the same as Chrome
