+++ browsers/chrome/manifest.json
@@ -40,7 +40,7 @@
         "matches": ["<all_urls>"],
         "all_frames": true,
         "css": ["public/css/noatb.css"],
-        "js": ["public/js/content-scripts/content-script-bundle.js", "public/js/inject-chrome.js"],
+        "js": ["public/js/content-scripts/content-script-bundle.js", "public/js/inject.js"],
         "run_at": "document_start"
       }
     ],
