+++ browsers/firefox/manifest.json
@@ -42,7 +42,7 @@
         "matches": ["<all_urls>"],
         "all_frames": true,
         "css": ["public/css/noatb.css"],
-        "js": ["public/js/content-scripts/content-script-bundle.js"],
+        "js": ["public/js/content-scripts/content-script-bundle.js", "public/js/inject.js"],
         "run_at": "document_start"
       }
     ],
