+++ shared/js/inject/chrome.js
@@ -61,6 +61,4 @@ function init() {
     );
 }
 
-
-
 init();
