+++ shared/js/inject/mozilla.js
@@ -1,0 +1,11 @@
+SCRIPT_TO_REPLACE
+
+function init() {
+    chrome.runtime.sendMessage({ registeredContentScript: true },
+      (message) => {
+          protections.initProtection(message)
+      }
+    );
+}
+
+init();
